# docs/backend/medcom-openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Medcom backend HTTP contract",
    "version": "2.0.0",
    "description": "All registered HTTP boundaries; full ERP business delivery is incomplete. Six v2 reads return all 116 qualified source columns. A route/schema never enables commands or proves target SQL/business acceptance."
  },
  "servers": [
    {
      "url": "/",
      "description": "This backend origin. FE uses its separately configured same-origin BFF."
    }
  ],
  "paths": {
...
    "/api/documents/field-contract": {
      "get": {
        "operationId": "documentFieldContract",
        "responses": {
          "200": {
            "description": "Typed response. Inspect access and outcome; HTTP 200 does not imply a committed write.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentFieldContract"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation, never client authority.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Medcom-Session-Scope": {
                "description": "Opaque current scope of a successful authorized read; invalidate stale data when it changes.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Medcom-Read-Scope": {
                "description": "Opaque current scope of a successful authorized read; invalidate stale data when it changes.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Non-success HTTP status. Do not infer an available operation or committed effect. Infrastructure errors may use another shape.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "x-medcom-admission": "module-read-capability-required",
        "description": "Contract of the registered boundary. Live credentials, capability, branch, session scope and provider qualification remain authoritative.",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "purchase-orders",
                "inbound-requests",
                "purchase-requests"
              ]
            }
          }
        ]
      }
    },
    "/api/documents/query-contract": {
      "get": {
        "operationId": "documentQueryContract",
        "responses": {
          "200": {
            "description": "Typed response. Inspect access and outcome; HTTP 200 does not imply a committed write.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentQueryContract"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation, never client authority.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Medcom-Session-Scope": {
                "description": "Opaque current scope of a successful authorized read; invalidate stale data when it changes.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Medcom-Read-Scope": {
                "description": "Opaque current scope of a successful authorized read; invalidate stale data when it changes.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Non-success HTTP status. Do not infer an available operation or committed effect. Infrastructure errors may use another shape.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "x-medcom-admission": "module-read-capability-required",
        "description": "Contract of the registered boundary. Live credentials, capability, branch, session scope and provider qualification remain authoritative.",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "purchase-orders",
                "inbound-requests",
                "purchase-requests"
              ]
            }
          }
        ]
      }
    },
    "/api/documents/purchase-orders": {
      "get": {
        "operationId": "purchaseordersLegacyList",
        "responses": {
          "200": {
            "description": "Typed response. Inspect access and outcome; HTTP 200 does not imply a committed write.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/purchaseordersLegacyPage"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation, never client authority.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Medcom-Session-Scope": {
                "description": "Opaque current scope of a successful authorized read; invalidate stale data when it changes.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Medcom-Read-Scope": {
                "description": "Opaque current scope of a successful authorized read; invalidate stale data when it changes.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Medcom-Data-Projection": {
                "description": "Successful document data projection. Full includes every qualified source column; summary retains the legacy wire shape.",
                "schema": {
                  "type": "string",
                  "const": "summary"
                }
              },
              "X-Medcom-Full-Data-Path": {
                "description": "Fixed complete-data route. Retain valid query parameters and adopt that route's full response schema.",
                "schema": {
                  "type": "string",
                  "const": "/api/v2/documents/purchase-orders"
                }
              }
            }
          },
          "default": {
            "description": "Non-success HTTP status. Do not infer an available operation or committed effect. Infrastructure errors may use another shape.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "x-medcom-admission": "qualified-read-provider-required",
        "description": "Contract of the registered boundary. Live credentials, capability, branch, session scope and provider qualification remain authoritative.",
        "x-medcom-data-projection": "summary",
        "x-medcom-full-data-path": "/api/v2/documents/purchase-orders",
        "x-medcom-capability": "purchase-orders.read",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "branchId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 50
            }
          }
        ]
      }
    },
    "/api/documents/purchase-orders/detail": {
      "get": {
        "operationId": "purchaseordersLegacyDetail",
        "responses": {
          "200": {
            "description": "Typed response. Inspect access and outcome; HTTP 200 does not imply a committed write.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/purchaseordersLegacyDetail"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation, never client authority.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Medcom-Session-Scope": {
                "description": "Opaque current scope of a successful authorized read; invalidate stale data when it changes.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Medcom-Read-Scope": {
                "description": "Opaque current scope of a successful authorized read; invalidate stale data when it changes.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Medcom-Data-Projection": {
                "description": "Successful document data projection. Full includes every qualified source column; summary retains the legacy wire shape.",
                "schema": {
                  "type": "string",
                  "const": "summary"
                }
              },
              "X-Medcom-Full-Data-Path": {
                "description": "Fixed complete-data route. Retain valid query parameters and adopt that route's full response schema.",
                "schema": {
                  "type": "string",
                  "const": "/api/v2/documents/purchase-orders/detail"
                }
              }
            }
          },
          "default": {
            "description": "Non-success HTTP status. Do not infer an available operation or committed effect. Infrastructure errors may use another shape.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "x-medcom-admission": "qualified-read-provider-required",
        "description": "Contract of the registered boundary. Live credentials, capability, branch, session scope and provider qualification remain authoritative.",
        "x-medcom-data-projection": "summary",
        "x-medcom-full-data-path": "/api/v2/documents/purchase-orders/detail",
        "x-medcom-capability": "purchase-orders.read",
        "parameters": [
          {
            "name": "documentId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 30
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ]
      }
    },
    "/api/v2/documents/purchase-orders": {
      "get": {
        "operationId": "purchaseordersV2List",
        "responses": {
          "200": {
            "description": "Typed response. Inspect access and outcome; HTTP 200 does not imply a committed write.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/purchaseordersV2Page"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation, never client authority.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Medcom-Session-Scope": {
                "description": "Opaque current scope of a successful authorized read; invalidate stale data when it changes.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Medcom-Read-Scope": {
                "description": "Opaque current scope of a successful authorized read; invalidate stale data when it changes.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Medcom-Data-Projection": {
                "description": "Successful document data projection. Full includes every qualified source column; summary retains the legacy wire shape.",
                "schema": {
                  "type": "string",
                  "const": "full"
                }
              },
              "X-Medcom-Full-Data-Path": {
                "description": "Fixed complete-data route. Retain valid query parameters and adopt that route's full response schema.",
                "schema": {
                  "type": "string",
                  "const": "/api/v2/documents/purchase-orders"
                }
              }
            }
          },
          "default": {
            "description": "Non-success HTTP status. Do not infer an available operation or committed effect. Infrastructure errors may use another shape.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "x-medcom-admission": "qualified-read-provider-required",
        "description": "Contract of the registered boundary. Live credentials, capability, branch, session scope and provider qualification remain authoritative.",
        "x-medcom-data-projection": "full",
        "x-medcom-full-data-path": "/api/v2/documents/purchase-orders",
        "x-medcom-capability": "purchase-orders.read",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "branchId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 50
            }
          },
          {
            "name": "dateFrom",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}$",
              "description": "SQL datetime calendar date, 1753-01-01 through 9999-12-31; no timezone conversion. dateFrom <= dateTo. dateTo includes its entire day."
            }
          },
          {
            "name": "dateTo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}$",
              "description": "SQL datetime calendar date, 1753-01-01 through 9999-12-31; no timezone conversion. dateFrom <= dateTo. dateTo includes its entire day."
            }
          },
          {
            "name": "statusId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": -2147483648,
              "maximum": 2147483647,
              "description": "Exact source StatusID. A filter never authorizes a state change."
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "documentDate",
                "documentId",
                "statusId"
              ],
              "default": "documentDate"
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc",
              "description": "Document ID uses binary byte ordering and breaks date/status ties ascending; SQL nulls sort first ascending and last descending. Offset pages are not a cross-request snapshot."
            }
          }
        ]
      }
    },
    "/api/v2/documents/purchase-orders/detail": {
      "get": {
        "operationId": "purchaseordersV2Detail",
        "responses": {
          "200": {
            "description": "Typed response. Inspect access and outcome; HTTP 200 does not imply a committed write.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/purchaseordersV2Detail"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation, never client authority.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Medcom-Session-Scope": {
                "description": "Opaque current scope of a successful authorized read; invalidate stale data when it changes.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Medcom-Read-Scope": {
                "description": "Opaque current scope of a successful authorized read; invalidate stale data when it changes.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Medcom-Data-Projection": {
                "description": "Successful document data projection. Full includes every qualified source column; summary retains the legacy wire shape.",
                "schema": {
                  "type": "string",
                  "const": "full"
                }
              },
              "X-Medcom-Full-Data-Path": {
                "description": "Fixed complete-data route. Retain valid query parameters and adopt that route's full response schema.",
                "schema": {
                  "type": "string",
                  "const": "/api/v2/documents/purchase-orders/detail"
                }
              }
            }
          },
          "default": {
            "description": "Non-success HTTP status. Do not infer an available operation or committed effect. Infrastructure errors may use another shape.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "x-medcom-admission": "qualified-read-provider-required",
        "description": "Contract of the registered boundary. Live credentials, capability, branch, session scope and provider qualification remain authoritative.",
        "x-medcom-data-projection": "full",
        "x-medcom-full-data-path": "/api/v2/documents/purchase-orders/detail",
        "x-medcom-capability": "purchase-orders.read",
        "parameters": [
          {
            "name": "documentId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 30
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          }
        },
        "required": []
      },
...
      "DocumentFieldContract": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "contractVersion": {
            "type": "integer",
            "format": "int32"
          },
          "kind": {
            "type": "string"
          },
          "header": {
            "$ref": "#/components/schemas/DocumentFieldTable"
          },
          "lines": {
            "$ref": "#/components/schemas/DocumentFieldTable"
          }
        },
        "required": [
          "contractVersion",
          "kind",
          "header",
          "lines"
        ]
      },
      "DocumentFieldTable": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "table": {
            "type": "string"
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentFieldDefinition"
            }
          }
        },
        "required": [
          "table",
          "fields"
        ]
      },
...
      "DocumentQueryContract": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "kind": {
            "type": "string"
          },
          "listPath": {
            "type": "string"
          },
          "maximumPage": {
            "type": "integer",
            "format": "int32"
          },
          "maximumPageSize": {
            "type": "integer",
            "format": "int32"
          },
          "dateColumn": {
            "type": "string"
          },
          "identifierColumn": {
            "type": "string"
          },
          "sortFields": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "defaultSortBy": {
            "type": "string"
          },
          "defaultSortDirection": {
            "type": "string"
          },
          "minimumDate": {
            "type": "string"
          },
          "maximumDate": {
            "type": "string"
          },
          "dateFormat": {
            "type": "string"
          },
          "dateToInclusive": {
            "type": "boolean"
          },
          "nullableDatesIncludedWithoutBounds": {
            "type": "boolean"
          },
          "statusColumn": {
            "type": "string"
          }
        },
        "required": [
          "version",
          "kind",
          "listPath",
          "maximumPage",
          "maximumPageSize",
          "dateColumn",
          "identifierColumn",
          "sortFields",
          "defaultSortBy",
          "defaultSortDirection",
          "minimumDate",
          "maximumDate",
          "dateFormat",
          "dateToInclusive",
          "nullableDatesIncludedWithoutBounds",
          "statusColumn"
        ]
      },
...
      "purchaseordersLegacyPage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/purchaseordersLegacySummary"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "rows",
          "page",
          "pageSize",
          "hasMore"
        ]
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-Medcom.Session"
      },
      "CsrfCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-Medcom.Csrf"
      },
      "CsrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-TOKEN"
      }
    }
  },
...
}